refactor: move utils functions to handlers package
Needed EnsureOwnerRefForSecret in the mutation handler.
